core/src/requirements.txt
fastapi==0.111.0
pytest==8.2.2
uvicorn==0.29.0
sqlalchemy==2.0.23
psycopg2==2.9.9
PyJWT==2.8.0
bcrypt==4.1.3
passlib==1.7.4
python-dotenv==1.0.1
itsdangerous==2.1.2
aiosmtplib==4.0.0
python-telegram-bot==21.10
requests==2.31.0